Split FileStreamWriter and MemoryStreamWriter into separate files
